requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
feedparser==6.0.11
anthropic==0.49.0
jinja2==3.1.4
python-dotenv==1.0.1
schedule==1.2.2
playwright==1.44.0
